+++ src/factorizations/factorizations.jl
@@ -6,7 +6,8 @@ module Factorizations
 export copy_oftype, factorisation_scalartype, one!, truncspace
 
 using ..TensorKit
-using ..TensorKit: AdjointTensorMap, SectorDict, SectorVector,
+using ..TensorKit: AdjointTensorMap, DictGradedSpace, SectorDict, SectorVector,
+    TupleGradedSpace,
     blocktype, foreachblock, one!,
     similar_diagonal, similarstoragetype
 using ..TensorKit: GLOBAL_TIMER